requirements.txt
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dj-rest-auth==7.0.1
django-cors-headers==4.4.0
boto3==1.34.131
reportlab==4.2.0
python-dotenv==1.0.1
dj-database-url==2.2.0
django-allauth==0.62.0
requests>=2.28.0
python-dotenv
